#reserva Facil

+++ src/br/com/rf/scheduling/entity/AbstractTransaction.java
@@ -1,14 +1,20 @@
 package br.com.rf.scheduling.entity;
 
+import java.math.BigDecimal;
+
+/**
+ * @author tiago.freitas
+ * Abstração do tipo de transação
+ */
 public abstract class AbstractTransaction implements Transaction{
 	
-	private double transferValue;
-
-	public AbstractTransaction(double transferValue) {
+	private BigDecimal transferValue;
+	
+	public AbstractTransaction(BigDecimal transferValue) {
 		this.transferValue = transferValue;
 	}
 
-	public double getTransferValue() {
+	public BigDecimal getTransferValue() {
 		return transferValue;
 	}
